Create now works
After not being able to get through on the curl requests, and solving
that with the help of mike, I was able to get into the back end adapter
of this and figure out what the data record was looking like, adapt
the way it calls the id and get it from there. I am psyched.